Varios ajustes
-Crear lógica para que todas las tablas que modifiquen alguna celda se modifique la celda correctamente sin regresar a su valor original
-En el modulo de devoluciones parciales por algún motivo cuando modifico la cantidad del producto a devolver no la cambia al perder el foco del teclado
-En el repositorio de productos agregar un método para consultar si una línea esta asociada a algún producto ya
-En el catálogo de líneas que no se permita borrar una línea si ya esta asociada a algún producto
-En el modulo de las líneas había un error de regresión que no permitía borrar la línea
-En el modulo de ventas que se abra la pantalla de devolución completa de ventas
-En el modulo de ventas que se abra la pantalla de devolución parcial de ventas
-Crear la clase EmailUtility escalable controladora gestora que es la que se comunica a nivel de envíos de correos electrónicos
-Integrar en maven la librería JavaMail API para envios de correos electrónicos con java
-En la pantalla de configuración de envios de correos electrónicos quitar los campos innecesarios
-En la pantalla de configurador de envios de correos electrónicos para los documentos y demás que ya funcione
-En la pantalla de configurador de envios de correos electrónicos cuando se hace la prueba de envio que muestre la forma de loading mientras el usuario espera al

proceso
-En el repositorio de correos agregue un método para consultar si un usuario ya tiene configuración de correo electrónico asociado
-Cuando se hacen cancelaciones de ventas no estaba afectando correctamente los inventarios cuando las cantidades eran mayores a 1
-Cuando se hacen devoluciones de ventas no estaba afectando correctamente los inventarios cuando las cantidades eran mayores a 1
-Cuando se hacen devoluciones parciales de ventas no estaba afectando correctamente los inventarios cuando las cantidades eran mayores a 1
-En el punto de venta cuando se va a agregar un producto que tiene precio 0 que mande una alerta de error
-En el catálogo de productos cuando se abre la lista de precios que coloque el cursor justamente en el precio 1
-En el punto de venta si al intentar agregar un kit este no tiene componentes que avise al usuario y que no permita continuar
-En el modulo de ventas que la pantalla de cancelar ventas se abra
-En el modulo de ventas rediseñar la pantalla existente de nuevas ventas para que este mas sencilla de utilizar y mas comodo a la vista
-En el modulo de ventas que las nuevas pantallas del nuevo diseño para las nuevas ventas funcionen correctamente todas en conjunto
-En el modulo de ventas al crear una nueva venta que el botón de ver descripción del producto en grande fucione
-En el modulo de ventas al crear una nueva venta que el botón de ver ver existencias de almacén del producto fucione
-En el modulo de ventas que ya se pueda generar una venta
-En el modulo de ventas que se pueda ver una venta al seleccionarla
-Crear un dialogo comun escalable para cuando la cantidad es incorrecta en donde pueda aplicar
-En el repositorio de productos crear un método para que devuelva si un número de lista tiene precio establecido o no
-En el repositorio de productos crear un método para que devuelva el precio de una lista
-En el modulo de ventas al crear una nueva venta al seleccionar una lista de precios de un producto y este no cuenta con precio para esa lista entonces que no permita

avanzar y que deje la lista anterior con precio
-Crear el combobox de monedas para que se pueda utilizar en todo el sistema
-Crear el combobox de método de pago para que se pueda utilizar en todo el sistema

+++ src/main/java/com/era/easyretail/controllers/views/LinesViewController.java
@@ -344,7 +344,7 @@ public class LinesViewController extends LinesJFrame {
                 DialogsFactory.getSingleton().showErrorFileNotExistsOKDialog(baseJFrame, (JFrame jFrame) -> {
                 });
                 return;
-            }
+            }                        
             
             //Question if continue
             DialogsFactory.getSingleton().showQuestionContinueDialog(baseJFrame, (JFrame jFrame) -> {
@@ -444,12 +444,21 @@ public class LinesViewController extends LinesJFrame {
                 return;
             }
             
+            //Get selected object
+            final Line Line = (Line)jTab.getRowSelected();
+                    
+            //Check if linea is associated to any product
+            final boolean existsLineInProduct = RepositoryFactory.getInstance().getProductsRepository().existsLineInProduct(Line.getCode());
+            if(existsLineInProduct){
+                DialogsFactory.getSingleton().showErrorOKCallbackDialog(baseJFrame, "errors_line_associated_with_product_stop", (JFrame jFrame) -> {
+                    jTab.grabFocus();
+                });
+               return; 
+            }
+            
             DialogsFactory.getSingleton().showQuestionContinueDialog(baseJFrame, (JFrame jFrame) -> {
                 
                 try {
-                    
-                    //Get selected object
-                    final Line Line = (Line)jTab.getRowSelected();
     
                     //Delete from the db
                     RepositoryFactory.getInstance().getLinesRepository().delete(Line);
